+++ buildout/src/Euphorie/euphorie/client/conditions.py
@@ -1,3 +1,4 @@
+from Acquisition import aq_inner
 from zope.interface import Interface
 from five import grok
 from AccessControl import getSecurityManager
@@ -5,6 +6,14 @@ from euphorie.client.interfaces import IClientSkinLayer
 from euphorie.client import CONDITIONS_VERSION
 
 grok.templatedir("templates")
+
+
+def approvedTermsAndConditions(account=None):
+    if account is None:
+        account=getSecurityManager().getUser()
+    return account.tc_approved is not None and account.tc_approved==CONDITIONS_VERSION
+
+
 
 class TermsAndConditions(grok.View):
     grok.name("terms-and-conditions")
@@ -18,7 +27,18 @@ class TermsAndConditions(grok.View):
 
 
     def update(self):
+        self.came_from=self.request.form.get("came_from")
+        if isinstance(self.came_from, list):
+            # If came_from is both in the querystring and the form data
+            self.came_from=self.came_from[0]
+
         self.account=getSecurityManager().getUser()
         if self.request.environ["REQUEST_METHOD"]=="POST":
             self.account.tc_approved=CONDITIONS_VERSION
+
+            if self.came_from:
+                self.request.response.redirect(self.came_from)
+            else:
+                self.request.response.redirect(
+                        aq_inner(self.context).absolute_url())